# perf: improve auto linking performance 