--- a/packages/cli-config-android/src/config/findComponentDescriptors.ts
+++ b/packages/cli-config-android/src/config/findComponentDescriptors.ts
@@ -5,7 +5,20 @@ import {extractComponentDescriptors} from './extractComponentDescriptors';
 import {unixifyPaths} from '@react-native-community/cli-tools';
 
 export function findComponentDescriptors(packageRoot: string) {
-  const files = glob.sync('**/+(*.js|*.jsx|*.ts|*.tsx)', {
+  let jsSrcsDir = null;
+  try {
+    const packageJson = fs.readFileSync(
+      path.join(packageRoot, 'package.json'),
+      'utf8',
+    );
+    jsSrcsDir = JSON.parse(packageJson).codegenConfig.jsSrcsDir;
+  } catch (error) {
+    // no jsSrcsDir, continue with default glob pattern
+  }
+  const globPattern = jsSrcsDir
+    ? `${jsSrcsDir}/**/+(*.js|*.jsx|*.ts|*.tsx)`
+    : '**/+(*.js|*.jsx|*.ts|*.tsx)';
+  const files = glob.sync(globPattern, {
     cwd: unixifyPaths(packageRoot),
     onlyFiles: true,
     ignore: ['**/node_modules/**'],