--- a/packages/cli-config-android/src/config/findPackageClassName.ts
+++ b/packages/cli-config-android/src/config/findPackageClassName.ts
@@ -25,6 +25,8 @@ export function getMainActivityFiles(
 
   return glob.sync(`**/+(${patternArray.join('|')})`, {
     cwd: unixifyPaths(folder),
+    onlyFiles: true,
+    ignore: ['**/.cxx/**'],
   });
 }