<!-- Thank you for sending the PR! We appreciate you spending the time to work on these changes. Please provide enough information so that others can review your pull request. -->

## Summary

There are multiple parts of auto linking that are slow, so I made a quick perf pass to reduce the time CLI spends on searching for native files. The biggest gain comes from optimizing `findComponentDescriptors` to use `jsSrcsDir`. For a small project with 7 common native dependencies the linking time goes from ~300ms to ~80ms. Perf and gains are bigger with more dependencies, so it's hard to asses the raw %.

|before|after|
|--|--|
|<img width="601" alt="Screenshot 2025-06-05 at 09 48 31" src="https://github.com/user-attachments/assets/e2a21b31-8a2d-4536-8c14-6e67c5c72bc1" />|<img width="610" alt="Screenshot 2025-06-05 at 09 49 29" src="https://github.com/user-attachments/assets/f67ef5cf-aee9-46e9-9337-cddcda89e0d9" />|

Changes are scoped to smaller commits:
* cdabe1cf - remote node_modules and pods from findManifest as we're in android folder
* 79c494fb - add android to exclude pattern in findAllPodfilePaths as it's not scoped to sourceDir 
* bbc597b6 - exclude .cxx from findManifest and findPackageClassName
* 503383f7 - use codegen config to find component descriptors
* d1e8f1aa - avoid double invoking of native deps - when calling `config`, the native dependencies getters are invoked twice: first for the filtering pass, and then for stringily. Changed the filtering logic to push resolved getters of native deps instead of removing non-native dependencies

## Test Plan

Make sure the projects are still properly linked

## Checklist

- [x] Documentation is up to date.
- [x] Follows commit message convention described in [CONTRIBUTING.md](https://github.com/react-native-community/cli/blob/main/CONTRIBUTING.md#commit-message-convention).
- [x] For functional changes, my test plan has linked these CLI changes into a local `react-native` checkout ([instructions](https://github.com/react-native-community/cli/blob/main/CONTRIBUTING.md#testing-your-changes)).